+++ src/Bzn.Cloudios.Domain/Enums/UserRole.cs
@@ -1,0 +1,12 @@
+namespace Bzn.Cloudios.Domain.Enums;
+
+public enum UserRole
+{
+    PlatformAdmin,
+    PlatformUser,
+    PlatformSre,
+    RealmOwner,
+    RealmAdmin,
+    RealmUser,
+    RealmSre
+}